refact: enum 분리 및 기타 리팩토링
